requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.5.0
--extra-index-url https://download.pytorch.org/whl/cu118
torchvision==0.20.0

opencv-python==4.10.0.84
pyyaml==6.0.2
matplotlib==3.10.0
einops==0.8.1